minor updates
added to hit_or_stay, removed the random card for the dealer since it now belongs to the Deck class and the Turn class. I'm also going to work on making an entire new file because this one is super confusing to me and I have no idea where things are pointing to anymore.